backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
discord.py==2.0.1
Django==4.1.1
django-configurations==2.4
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-extensions==3.2.1
django-filter==22.1
django-graphql-jwt==0.3.4
django-import-export==2.9.0
django-markdownx==4.0.0b1
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
frozenlist==1.3.1
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.3
importlib-metadata==5.0.0
Markdown==3.4.1
MarkupPy==1.14
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
promise==2.3
psycopg2==2.9.3
pycodestyle==2.9.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.12.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
types-pytz==2022.4.0.0
types-PyYAML==6.0.12
types-requests==2.28.10
types-urllib3==1.26.24
typing_extensions==4.4.0
Unidecode==1.3.4
urllib3==1.26.12
uWSGI==2.0.20
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.1
zipp==3.8.1
